requirements.txt
Django==1.6.4
django-configurations==0.8
dj-database-url==0.3.0
pylibmc==1.3.0
Pillow==2.0.0
django-cache-url==0.8.0
werkzeug==0.9.4
gunicorn==0.17.4
easy-thumbnails==1.2
django-extensions==1.3.4
django-braces==1.4.0
django-allauth==0.16.1
django-zurb-foundation==5.1.1
django-floppyforms==1.1.1
django-custom-user==0.4
django-localflavor==1.0
django-taggit==0.12
south==0.8.2

boto==2.9.5
django-storages==1.1.8

psycopg2==2.5
django-subdomains==2.0.4
